refactor: seperate components, types and util functions from main page
