+++ frontend/components/common/Loading.tsx
@@ -1,0 +1,5 @@
+export const Loading = () => (
+  <div className="inline-block h-8 w-8 animate-spin rounded-full border-4 border-gray-200 border-t-cyan-400 border-r-pink-200/30 border-b-[#332c30]">
+    <span className="sr-only">Loading...</span>
+  </div>
+);
