+++ frontend/lib/utils.ts
@@ -1,0 +1,3 @@
+export const getFaviconUrl = (url: string): string => {
+  return `https://www.google.com/s2/favicons?domain=${url}&sz=64`;
+};
