+++ frontend/types/linkCard.ts
@@ -1,0 +1,5 @@
+export interface LinkCardProps {
+  id: number;
+  name: string;
+  url: string;
+}
